tutorial/requirements.txt
Django==1.9.12
djangorestframework
markdown
django-filter
pygments
httpie
coreapi
coreapi-cli
django-rest-swagger
mysql-python>=1.2.5